Move remaining network processing into LoginServer class
