+++ game/src/main/kotlin/world/gregs/voidps/Main.kt
@@ -22,6 +22,7 @@ import world.gregs.voidps.engine.client.update.iterator.SequentialIterator
 import world.gregs.voidps.engine.data.definition.*
 import world.gregs.voidps.engine.entity.World
 import world.gregs.voidps.engine.map.collision.CollisionDecoder
+import world.gregs.voidps.network.LoginServer
 import world.gregs.voidps.network.Network
 import world.gregs.voidps.network.protocol
 import world.gregs.voidps.script.loadScripts
@@ -49,7 +50,10 @@ object Main {
 
         val accountLoader = PlayerAccountLoader(get<ConnectionQueue>(), get(), Contexts.Game)
         val protocol = protocol(get<Huffman>())
-        val server = Network.load(cache, properties, protocol, get<ConnectionGatekeeper>(), accountLoader, Contexts.Game)
+
+        val gatekeeper: ConnectionGatekeeper = get()
+        val loginServer = LoginServer.load(properties, protocol, gatekeeper, accountLoader, Contexts.Game)
+        val server = Network.load(cache, properties, gatekeeper, loginServer)
 
         val tickStages = getTickStages(iterator = if (CharacterTask.DEBUG) SequentialIterator() else ParallelIterator())
         val engine = GameLoop(tickStages)
