+++ network/src/main/kotlin/world/gregs/voidps/network/Request.kt
@@ -1,0 +1,16 @@
+package world.gregs.voidps.network
+
+object Request {
+    const val PREFETCH_REQUEST = 0
+    const val PRIORITY_REQUEST = 1
+    const val STATUS_LOGGED_IN = 2
+    const val STATUS_LOGGED_OUT = 3
+    const val ENCRYPTION_KEY_UPDATE = 4
+    const val ACKNOWLEDGE = 6
+    const val SESSION = 10
+    const val CONNECT_LOGIN = 14
+    const val CONNECT_JS5 = 15
+    const val LOGIN = 16
+    const val RECONNECT = 18
+    const val SIGN_UP = 28
+}
