+++ network/src/main/kotlin/world/gregs/voidps/network/Server.kt
@@ -1,0 +1,7 @@
+package world.gregs.voidps.network
+
+import io.ktor.utils.io.*
+
+interface Server {
+    suspend fun connect(read: ByteReadChannel, write: ByteWriteChannel, hostname: String)
+}
